Backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
pymongo==4.6.1
python-dotenv==1.0.0
Werkzeug==3.0.1

# WebSocket for real-time
flask-socketio==5.3.6
python-socketio==5.11.1
gevent==24.2.1
gevent-websocket==0.10.1

# AI/ML features for transcription and summarization
google-generativeai>=0.3.0  # Google Gemini (free tier - 15 RPM)
groq>=0.4.0  # Groq (free tier - extremely fast, recommended fallback)
openai>=1.99.9  # OpenAI (paid - optional fallback)

# PDF generation for session reports
reportlab==4.0.9

# Video Emotion Detection (100% free, local inference)
# Tested working configuration as of Dec 2024
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
mediapipe>=0.10.0
fer>=22.5.0
tensorflow>=2.13.0,<2.18.0  # TensorFlow 2.17.1 confirmed working
numpy>=1.24.0,<2.0.0  # numpy 1.26.4 compatible with langchain
torch>=2.0.0  # PyTorch for facenet-pytorch (FER dependency)

# Note: jax and jaxlib removed to avoid conflicts (not needed for emotion detection)
